Changed analogRead to use bone.PIN for pin #
You can just pass the AINx number via {'ain':x} as that is currently the
only parameter utilized in the pin structure for now.  The benefit is to
keep the code written by the user to still document the hardware pins in
a way that doesn't require looking something up in a document.