testing: used burrito to parse down to issue
Current issue is that variables defined in the main part of the program
aren't visible to the forks.  I'm a bit confused by this as I thought
they'd be redefined in the forks, but it is good to figure out how to
resolve them now such that loop functions passed to other processors
have a chance to run.