Create a simulator mode
Move all hardware accesses into one of 3 files:
* hw_capemgr.js: for the 3.8.x kernel with capemgr
* hw_oldkernel.js: for the 3.2.x kernel with omapmux
* hw_simulator.js: for other platforms

Goal of this is to try to produce a test bench for some basic logic that
can be run on any hardware. Also, the idea is to make things easier when
it is time to update to the 3.12/3.13 kernel without hurting
compatibility.